requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.28.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
transformers==4.35.2
pillow==10.0.1
httpx==0.24.1
python-multipart==0.0.6
openai==1.3.7
tensorflow==2.12.0
numpy==1.23.5
scikit-learn==1.3.0
pandas==2.0.3
gunicorn==21.2.0
setuptools==68.2.2
wheel==0.41.2
requests==2.31.0
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
supabase==2.8.0
sendgrid==6.11.0
